move big functions from .h to .cpp
